Move calibrate function out of __init__ file
